+++ src/fields/MultipleSelect.js
@@ -22,8 +22,6 @@ class Select extends React.Component {
             items,
             ...otherProps} = this.props;
 
-        if (value === null) value = [];
-
         return <div className={validationState ? ('has-'+validationState):''}>
             <select {...{
                 className: "form-control",
@@ -48,5 +46,10 @@ class Select extends React.Component {
 
 Select.propTypes = propTypes;
 Select.defaultProps = defaultProps;
+Select.cleanValue = (value, {items}) => {
+    if (value === undefined) return value;
+    else if (!Array.isArray(value)) return [];
+    else return value.filter(key=>!!items[key])
+};
 
 export default Select;