+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+## 0.2.0
+- add cleanValue to all fields
+- fix missing add button of array
+
 ## 0.1.0
 first release